Add support for subcommands
When using creating a console application and using DocOpt, the most
common use case is to implement subcommands.

This patch provides a skeleton to achieve this. The ``commands`` module
comes with a base class for the commands, with the ability to set up
the python logger and to display information on the console through the
``tabulate`` library.

The ``q`` library was added to the test-requirements to ease the testing
of the application if needed.